+++ src/App.css
@@ -1,5 +1,19 @@
+@import "./reset.css";
 :root {
   --card-background: #333333;
   --card-chip: #ddcd78;
   --color-white: #ffffff;
+  --color-description: #8b95a1;
+  --color-placeholder: #acacac;
+}
+
+.App {
+  width: 376px;
+  justify-self: center;
+  display: flex;
+  flex-direction: column;
+}
+
+.card-input {
+  justify-self: center;
 }
