+++ src/App.tsx
@@ -1,17 +1,19 @@
 import "./App.css";
-import CardNumber from "./CardNumber";
-import CardExpirationDate from "./CardExpirationDate";
-import CardCvcNumber from "./CardCvcNumber";
+import CardNumber from "./CardNumber/CardNumber";
+import CardExpirationDate from "./CardExpirationDate/CardExpirationDate";
+import CardCvcNumber from "./CardCvcNumber/CardCvcNumber";
 import CardLayout from "./components/Card/CardLayout";
 
 function App() {
   return (
-    <>
+    <div className="App">
       <CardLayout />
-      <CardNumber />
-      <CardExpirationDate />
-      <CardCvcNumber />
-    </>
+      <div className="card-input">
+        <CardNumber />
+        <CardExpirationDate />
+        <CardCvcNumber />
+      </div>
+    </div>
   );
 }
