+++ src/CardCvcNumber/CardCvcNumber.module.css
@@ -1,0 +1,3 @@
+.card-cvc {
+  width: 100%;
+}
