+++ src/CardExpirationDate/CardExpirationDate.module.css
@@ -1,0 +1,5 @@
+.card-number__input {
+  display: flex;
+  flex-direction: row;
+  gap: 0.5rem;
+}
