+++ src/CardExpirationDate/CardExpirationDate.tsx
@@ -1,7 +1,8 @@
-import InputNumber from "./components/Input/InputNumber";
-import InputDescription from "./components/InputDescription/InputDescription";
-import InputSubtitle from "./components/InputSubtitle/InputSubtitle";
-import InputTitle from "./components/InputTitle/InputTitle";
+import InputNumber from "../components/Input/InputNumber";
+import InputDescription from "../components/InputDescription/InputDescription";
+import InputSubtitle from "../components/InputSubtitle/InputSubtitle";
+import InputTitle from "../components/InputTitle/InputTitle";
+import styles from "./CardExpirationDate.module.css";
 
 export default function CardExpirationDate() {
   return (
@@ -11,7 +12,7 @@ export default function CardExpirationDate() {
         inputValue={"월/년도(MMYY)를 순서대로 입력해 주세요."}
       />
       <InputSubtitle inputValue={"유효기간"} />
-      <div className="card-number__input">
+      <div className={styles["card-number__input"]}>
         <InputNumber />
         <InputNumber />
       </div>
