+++ src/components/Card/CardLayout.module.css
@@ -4,6 +4,7 @@
   background-color: var(--card-background);
   border-radius: 4px;
   position: relative;
+  margin: 77px 82px;
 }
 
 .card-chip {
