+++ src/components/Card/CardLayout.tsx
@@ -1,11 +1,11 @@
-import "./CardLayout.css";
+import styles from "./CardLayout.module.css";
 import CardLogo from "./CardLogo";
 import CardNumber from "./CardNumber";
 
 export default function CardLayout() {
   return (
-    <div className="card-layout">
-      <div className="card-chip"></div>
+    <div className={styles["card-layout"]}>
+      <div className={styles["card-chip"]}></div>
       <CardLogo />
       <CardNumber />
     </div>
