+++ src/components/Card/CardLogo.tsx
@@ -1,10 +1,14 @@
-import "./CardLogo.css";
+import styles from "./CardLogo.module.css";
 
 export default function CardLogo() {
   return (
-    <div className="card-logo">
+    <div className={styles["card-logo"]}>
       {/* TODO : 이미지 경로 props로 수정 */}
-      <img className="card-logo-image" src="./master-logo.png" alt="로고"></img>
+      <img
+        className={styles["card-logo-image"]}
+        src="./master-logo.png"
+        alt="로고"
+      ></img>
     </div>
   );
 }
