+++ src/components/Input/InputNumber.css
@@ -1,0 +1,3 @@
+.input-number {
+  width: 100%;
+}
