+++ src/components/InputDescription/InputDescription.css
@@ -1,0 +1,5 @@
+.input-description {
+  font-size: 9.5px;
+  font-weight: 500;
+  color: var(--color-description);
+}
