+++ src/components/InputSubtitle/InputSubtitle.css
@@ -1,0 +1,5 @@
+.input-subtitle {
+  font-size: 12px;
+  font-weight: 500;
+  padding: 6px 0;
+}
