+++ src/components/InputSubtitle/InputSubtitle.tsx
@@ -1,3 +1,4 @@
+import "./InputSubtitle.css";
 interface Props {
   inputValue: string;
 }
