+++ src/components/InputTitle/InputTitle.css
@@ -1,0 +1,5 @@
+.input-title {
+  font-size: 18px;
+  font-weight: 700;
+  padding: 6px 0;
+}
